requirements.txt
streamlit
tensorflow
numpy
opencv-python
Pillow
gTTS
boto3
google-generativeai